Server/attinit/requirements.txt
wheel
click
mysql.connector
